+++ src/admin/js/plugins/publish/component.tsx
