+++ src/admin/js/sanity.ts
@@ -3,7 +3,7 @@ import { defineConfig, defineType, renderStudio } from "sanity";
 import { UserIcon } from "@sanity/icons";
 import { media } from "sanity-plugin-media";
 import { admin } from "../../../photo-grid.json";
-import { deploy } from "./plugins/publish";
+import { publish } from "./plugins/publish";
 
 const config = defineConfig({
   basePath: "/admin",
@@ -12,7 +12,7 @@ const config = defineConfig({
     newDocumentOptions: () => [],
   },
   icon: UserIcon,
-  plugins: [media(), deploy()],
+  plugins: [media(), publish()],
   projectId: import.meta.env.SANITY_STUDIO_PROJECT_ID,
   title: admin.title,
   schema: {
